Revert "Move the car project in one file"
This reverts commit a43b1ac592a835939ee2af77dee7b992c5ca3340.
