+++ basic-2/src/test/java/Bicycle.java
@@ -1,0 +1,17 @@
+import java.lang.invoke.ConstantBootstraps;
+
+public class Bicycle extends Vehicle{
+	private int odometer=0;
+
+	public int getOdometer() {
+		return this.odometer;
+	}
+	
+	public void drive(int distance) {
+		this.odometer+=distance;
+	}
+	
+	boolean isDriveable() {
+		return true;
+	}
+}
