+++ basic-2/src/test/java/IEngine.java
@@ -1,0 +1,6 @@
+
+public interface IEngine {
+	
+	public int getCylinderCount();
+	public void start();
+}
