+++ basic-2/src/test/java/LargeEngine.java
@@ -1,0 +1,13 @@
+
+public class LargeEngine implements IEngine{
+	
+	public int cylinderCount=8;
+	
+	public void start() {
+		System.out.println("Engine Started");
+	}	
+	
+	public int getCylinderCount() {
+		return this.cylinderCount;
+	}
+}
