+++ basic-2/src/test/java/Vehicle.java
@@ -1,0 +1,8 @@
+
+public abstract class Vehicle {
+	
+	public String model;
+	
+	abstract int getOdometer();
+	abstract void drive(int distance);
+}
