--- basic-2/src/test/java/cars/Bicycle.java
@@ -1,17 +1,0 @@
-package cars;
-
-public class Bicycle extends Vehicle{
-	private int odometer=0;
-
-	public int getOdometer() {
-		return this.odometer;
-	}
-	
-	public void drive(int distance) {
-		this.odometer+=distance;
-	}
-	
-	boolean isDriveable() {
-		return true;
-	}
-}
