--- basic-2/src/test/java/cars/IEngine.java
@@ -1,7 +1,0 @@
-package cars;
-
-public interface IEngine {
-	
-	public int getCylinderCount();
-	public void start();
-}
