--- basic-2/src/test/java/cars/LargeEngine.java
@@ -1,14 +1,0 @@
-package cars;
-
-public class LargeEngine implements IEngine{
-	
-	public int cylinderCount=8;
-	
-	public void start() {
-		System.out.println("Engine Started");
-	}	
-	
-	public int getCylinderCount() {
-		return this.cylinderCount;
-	}
-}
