--- basic-2/src/test/java/cars/SmallEngine.java
@@ -1,14 +1,0 @@
-package cars;
-
-public class SmallEngine implements IEngine{
-	
-	public int cylinderCount=4;
-	
-	public void start() {
-		System.out.println("Engine Started");
-	}	
-	
-	public int getCylinderCount() {
-		return this.cylinderCount;
-	}
-}
