--- basic-2/src/test/java/cars/Vehicle.java
@@ -1,9 +1,0 @@
-package cars;
-
-public abstract class Vehicle {
-	
-	public String model;
-	
-	abstract int getOdometer();
-	abstract void drive(int distance);
-}
